Data packing attempt 2 Fix thing with sign bits

+++ labs/memory_bound/data_packing/solution.h
@@ -9,7 +9,7 @@ constexpr int maxRandom = 100;
 struct S {
   unsigned long long l : 14;
   unsigned i : 7;
-  short s : 7;
+  unsigned s : 7;
   float d;
   bool b;
